[CHANGED] src[8]改成用float
Former-commit-id: 495317c4133d81b3364c691c11ba3c710164e526

+++ ejoy2d/src/eploader.h
@@ -10,7 +10,8 @@ extern "C"
 
 struct picture_part {
   int32_t texid;
-  int16_t src[8];
+//  int16_t src[8];
+  float src[8];
   int32_t screen[8];
 };
